pyproject.toml
[tool.poetry]
name = "scribe"
version = "0.1.0"
description = "Audio transcription and text summarization service"
authors = ["t-ashula"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
pydantic = "^2.6.0"
python-dotenv = "^1.0.0"
httpx = "^0.28.1"
python-multipart = "^0.0.20"
redis = "^5.0.1"
rq = "^1.15.1"
rq-scheduler = "^0.13.1"
python-magic = "^0.4.27"
python-ulid = "^3.0.0"
transformers = "^4.40.2"
punctuators = "0.0.5"
accelerate = "^0.30.0"
torchaudio = "^2.3.0"
stable-ts = "2.16.0"
datasets = "^2.19.1"
librosa = "^0.10.2"
soundfile = "^0.12.1"
torch = "^2.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.8.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88
